+++ packages/react/stories/greyVest/toggleFiltersButton.stories.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import { action } from '@storybook/addon-actions'
+import decorator from './decorator'
+import { ToggleFiltersButton } from './../../src/themes/greyVest'
+
+let click = action('clicked')
+
+storiesOf('Components (Grey Vest)|ToggleFiltersButton', module)
+  .addDecorator(decorator)
+  .addWithJSX('Basic Usage', () => (
+    <ToggleFiltersButton onClick={() => click()} />
+  )
+)